+++ nata-monkey/src/main/java/com/nata/monkeys/RandomMonkey.java
@@ -1,10 +1,9 @@
 package com.nata.monkeys;
 
-import com.nata.action.Action;
-import com.nata.action.BackAction;
-import com.nata.action.TapAction;
+import com.nata.action.*;
 import com.nata.cmd.AdbDevice;
 import com.nata.element.Element;
+import com.nata.element.UINode;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -17,9 +16,26 @@ import java.util.List;
 public class RandomMonkey extends AbstractMonkey {
     volatile private boolean isRunning = true;
     private List<Action> actionList = new ArrayList<>();
+    private final int ACTION_LIMIT = 20;
 
     public RandomMonkey(String pkg, String act, AdbDevice device){
         super("randomMonkey",pkg,act,device);
+    }
+
+
+    private class UINodeAction{
+        UINode node;
+        String Action;
+        UINodeAction(UINode node,String Action){
+            this.node = node;
+            this.Action = Action;
+        }
+        public UINode getNode(){
+            return node;
+        }
+        public String getAction(){
+            return Action;
+        }
     }
 
     @Override
@@ -27,18 +43,47 @@ public class RandomMonkey extends AbstractMonkey {
         System.out.println("start playing...");
         startApp();
         while(isRunning){
-            Element nextElement = getNextElement();
-            Action nextAction = null;
-            if(nextElement!=  null){
-                nextAction = new TapAction(nextElement,getDevice());
-            }else {
+            Action nextAction = getNextUIAction();
+            if(nextAction==  null){
                 nextAction = new BackAction(getDevice());
             }
             actionList.add(nextAction);
             nextAction.fire();
             System.out.println(nextAction);
+            if(actionList.size() >= ACTION_LIMIT){
+               break;
+            }
         }
         report();
+    }
+
+
+    public Action getNextUIAction(){
+        List<UINode> list = GrabCurrentUi();
+        List<UINodeAction> actionList  = new ArrayList<>();
+        if(list != null){
+            for (UINode node: list) {
+                if(node.getClassName().equals("android.widget.Button") && node.getClickable().equals("true")){
+                    actionList.add(new UINodeAction(node, ActionType.TAP));
+                }
+                if(node.getClassName().equals("android.widget.EditText")){
+                    actionList.add(new UINodeAction(node, ActionType.INPUT));
+                }
+            }
+            if(actionList.size() > 0){
+                int randomIndex = (int)(Math.random() * actionList.size());
+                UINodeAction nodeAction = actionList.get(randomIndex);
+                Action action = null ;
+                switch (nodeAction.getAction()){
+                    case  ActionType.INPUT :action = new TextInputAction(new Element(nodeAction.getNode().getBounds()),getDevice());break;
+                    case  ActionType.TAP :action = new TapAction(new Element(nodeAction.getNode().getBounds()),getDevice());break;
+                }
+                return  action;
+            }else {
+                return null;
+            }
+        }
+        return null;
     }
 
     @Override
@@ -48,13 +93,15 @@ public class RandomMonkey extends AbstractMonkey {
 
     @Override
     public void report() {
-        System.out.println("report");
+        System.out.println(actionList);
     }
 
     public static void main(String[] args) {
-        String pkg = "com.zhihu.android";
-        String act = ".app.ui.activity.MainActivity";
+//        String pkg = "com.zhihu.android";
+//        String act = ".app.ui.activity.MainActivity";
 
+        String pkg = "random";
+        String act = "random";
         AdbDevice device = new AdbDevice();
         RandomMonkey randomMonkey = new RandomMonkey(pkg,act,device);